Replaced the rebuild renderers prowindow with a tab
