+++ VSProjects/Geomapmaker/ViewModels/Tools/ToolsViewModel.cs
@@ -2,21 +2,39 @@
 using ArcGIS.Desktop.Framework.Contracts;
 using ArcGIS.Desktop.Framework.Controls;
 using System;
+using System.ComponentModel;
+using System.Runtime.CompilerServices;
 using System.Windows.Input;
 
 namespace Geomapmaker.ViewModels.Tools
 {
-    public class ToolsViewModel : ProWindow
+    public class ToolsViewModel : ProWindow, INotifyPropertyChanged
     {
         public event EventHandler WindowCloseEvent;
 
         public ICommand CommandCancel => new RelayCommand(() => CloseProwindow());
+
+        public SymbologyToolsVM SymbologyTools { get; set; }
+
+        public ToolsViewModel()
+        {
+            SymbologyTools = new SymbologyToolsVM(this);
+        }
 
         public void CloseProwindow()
         {
             WindowCloseEvent(this, new EventArgs());
         }
 
+        #region INotifyPropertyChanged
+
+        public event PropertyChangedEventHandler PropertyChanged;
+
+        private void NotifyPropertyChanged([CallerMemberName] string propertyName = "")
+        {
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+        }
+        #endregion
     }
 
     internal class ShowTools : Button
