--- VSProjects/Geomapmaker/Views/RebuildRenderers/RebuildRenderers.xaml
@@ -1,51 +1,0 @@
-﻿<controls:ProWindow x:Class="Geomapmaker.Views.RebuildRenderers.RebuildRenderers"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:controls="clr-namespace:ArcGIS.Desktop.Framework.Controls;assembly=ArcGIS.Desktop.Framework"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:extensions="clr-namespace:ArcGIS.Desktop.Extensions;assembly=ArcGIS.Desktop.Extensions"
-        mc:Ignorable="d"
-        Title="Rebuild Renderers" 
-        Height="350" 
-        MinHeight="300" 
-        Width="500"
-        MinWidth="450"                     
-        WindowStartupLocation="CenterOwner"
-    >
-    <controls:ProWindow.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <extensions:DesignOnlyResourceDictionary Source="pack://application:,,,/ArcGIS.Desktop.Framework;component\Themes\Default.xaml"/>
-                <ResourceDictionary Source="/Geomapmaker;component/Views/GlobalResourceDictionary.xaml" />
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </controls:ProWindow.Resources>
-    
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="5"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="5"/>
-        </Grid.RowDefinitions>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="5"/>
-            <ColumnDefinition Width="*" />
-            <ColumnDefinition Width="5"/>
-        </Grid.ColumnDefinitions>
-
-        <Button Grid.Column="1" Grid.Row="1" Command="{Binding CommandAll}" Style="{DynamicResource Esri_Button}" Background="#2082D8" Foreground="#FFF" MinHeight="50" >Rebuild All</Button>
-        <Button Grid.Column="1" Grid.Row="2" Command="{Binding CommandCIMSymbols}" Style="{DynamicResource Esri_Button}" >Line/Point Symbol Options</Button>
-        <Button Grid.Column="1" Grid.Row="3" Command="{Binding CommandMups}" Style="{DynamicResource Esri_Button}" >MapUnitPolys Symbology and Templates</Button>
-        <Button Grid.Column="1" Grid.Row="4" Command="{Binding CommandCf}" Style="{DynamicResource Esri_Button}" >Contacts Faults Symbology</Button>
-        <Button Grid.Column="1" Grid.Row="5" Command="{Binding CommandOp}" Style="{DynamicResource Esri_Button}" >Orientation Points Symbology</Button>
-
-        <Button Grid.Column="1" Grid.Row="7" Command="{Binding CommandCancel}" Style="{DynamicResource Esri_Button}" MaxWidth="100" >Cancel</Button>
-    </Grid>
-</controls:ProWindow>
